Fix NPCBot arena gating and AV objective routing

+++ src/server/game/AI/NpcBots/bot_ai.h
@@ -767,7 +767,7 @@ private:
     uint32 _wandererZoneActivityTimer{};
     bool _wandererZoneHasPlayer{ true };
     mutable uint8 _avObjectiveRole{};
-    mutable uint32 _avObjectiveRoleExpireTime{};
+    mutable uint64 _avObjectiveRoleExpireTime{};
 
     bool IsWandererZoneInactive() const { return IsWanderer() && !_wandererZoneHasPlayer; }
